+++ next.config.js
@@ -1,9 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
-  images: {
-    domains: ['www.google.com', 'bit.ly', "images.unsplash.com", "c1.staticflickr.com", "c2.staticflickr.com", "c7.staticflickr.com", "c3.staticflickr.com", "c4.staticflickr.com", "c5.staticflickr.com", "c6.staticflickr.com", "c7.staticflickr.com", "c8.staticflickr.com", "c9.staticflickr.com",]
-  }
-}
+    reactStrictMode: true,
+    images: {
+        domains: ["images.unsplash.com"],
+    },
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
